+++ app.py
@@ -48,7 +48,7 @@ from utils import get_settings, utc_now
 from routes import register_all
 from messages import TRANSLATIONS
 
-APP_VERSION = '1.76.2'
+APP_VERSION = '1.77.0'
 
 csrf = CSRFProtect()